pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-bitget-trader"
version = "0.1.0"
description = "Telegram signal listener -> risk filter -> Bitget futures executor"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Trading System" }]
dependencies = [
  "pydantic>=2.6",
  "PyYAML>=6.0",
  "telethon>=1.35",
  "requests>=2.31",
  "beautifulsoup4>=4.12",
  "openai>=1.51.0",
  "typer>=0.12",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
trader = "trader.main:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["trader*"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
